backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.36,<2.1.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
pydantic>=2.11.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
httpx>=0.27.0
PyJWT>=2.8.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0,<5
